Use OAuth for authentication and authorization, Retrofit for REST.
The OAuth is defined by a bunch of services and implemented by making
use of the AppAuth library.

Retrofit is used for making REST request and responses are emitted on
Rx observables. The OkHttpClient used by Retrofit makes use of the OAuth
services to obtain and attach the correct access-tokens needed for the
HTTP requests that need authorization.

For now, there is only one dummy REST service called `TestRestService`.
It is used in the `HomeActivity` by logging the results of an SOQL query.