Add custom primitive types that are handled by our JSON serialization
The Date and Time classes are from the ThreeThenBp library.

LocalDate:
Stores a date without a time. This stores a date like '2010-12-03' and could be used to store a birthday.

LocalTime:
Stores a time of day without a date. This stores a time like '11:30' and could be used to store an opening or closing time.

LocalDateTime:
Stores a date and time of day. This stores a date-time like '2010-12-03T11:30'.

ZonedDateTime:
Stores a date and time with a time-zone.

Duration:
Stores a duration as a simple number of milliseconds.

CurrencyAmount:
Stores an amount of money of a given currency (defaults to USD)

EmailAddress:
Stores an properly constructed email address.

PhoneNumber:
Stores a US phone-number.